refactor: set up member in BaseIntegrationTest.
